+++ src/com/walking/lesson1_varaibles/Task2ComplicatedWithString.java
@@ -1,4 +1,4 @@
-package com.walking.lesson1;
+package com.walking.lesson1_varaibles;
 
 /**
  * Усложнение для внимательных: вывести ваше имя посимвольно,
