+++ src/com/walking/lesson1_varaibles/Task2SingleChar.java
@@ -1,4 +1,4 @@
-package com.walking.lesson1;
+package com.walking.lesson1_varaibles;
 
 /**
  * Используя одну переменную типа char вывести свое имя в консоль.
