+++ src/com/walking/lesson1_varaibles/Task2String.java
@@ -1,4 +1,4 @@
-package com.walking.lesson1;
+package com.walking.lesson1_varaibles;
 
 /**
  * Используя Переменную типа String вывести свое имя в консоль.
