+++ src/com/walking/lesson3_casts_conditional_constructions/Task1.java
@@ -1,4 +1,4 @@
-package com.walking.lesson3;
+package com.walking.lesson3_casts_conditional_constructions;
 
 import java.util.Scanner;
